Changed sout to be old class, created enum with program options.
