+++ VacationPlanner-ConsoleApp/src/main/java/com/isa/menu/MenuList_old.java
@@ -1,0 +1,19 @@
+package com.isa.menu;
+
+public class MenuList_old {
+
+    public void printOption() {
+
+        System.out.println("1 - Dodaj dzień urlopowy.");
+        System.out.println("2 - Dodaj pracownika.");
+        System.out.println("3 - Usuń pracownika.");
+        System.out.println("4 - Dodaj zespół.");
+        System.out.println("5 - Usuń zespół.");
+        System.out.println("0 - Wyjdź z programu");
+        System.out.println("\nWybierz opcje: ");
+    }
+
+
+}
+
+
